requirements-railway.txt
flask
easyocr
torch
pdf2image
Pillow
scikit-image
gunicorn
Werkzeug
